+++ cookietemple/lint/domains/cli.py
@@ -7,14 +7,17 @@ from pkg_resources import parse_version
 from rich import print
 
 from cookietemple.custom_cli.questionary import cookietemple_questionary_or_dot_cookietemple
-from cookietemple.lint.template_linter import TemplateLinter, files_exist_linting, GetLintingFunctionsMeta
+from cookietemple.lint.template_linter import TemplateLinter, files_exist_linting, GetLintingFunctionsMeta, ConfigLinter
 
 CWD = os.getcwd()
 
 
 class CliPythonLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
+
     def __init__(self, path):
         super().__init__(path)
+        self.blacklisted_sync_files = [('requirements', 'requirements.txt'), ('requirements_dev', 'requirements_dev.txt'), ('changelog', 'CHANGELOG.rst')]
+        self.blacklisted_lint_code = 'cli-python-3'
 
     def lint(self, is_create, skip_external):
         super().lint_project(self, self.methods)
@@ -34,6 +37,22 @@ class CliPythonLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
             autopep8 = Popen(['autopep8', self.path, '--recursive', '--in-place', '--pep8-passes', '2000'],
                              universal_newlines=True, shell=False, close_fds=True)
             (autopep8_stdout, autopep8_stderr) = autopep8.communicate()
+
+    def check_sync_section(self) -> bool:
+        """
+        Check the sync_files_blacklisted section containing every required file!
+        """
+        config_linter = ConfigLinter(f'{self.path}/cookietemple.cfg', self)
+        result = config_linter.check_section(section_items=config_linter.parser.items('sync_files_blacklisted'), section_name='sync_files_blacklisted',
+                                             main_linter=self,
+                                             blacklisted_sync_files=[[('requirements', 'requirements.txt'), ('requirements_dev', 'requirements_dev.txt'),
+                                                                      ('changelog', 'CHANGELOG.rst')], -1], error_code='cli-python-3',
+                                             is_sublinter_calling=True)
+        if result:
+            self.passed.append(('cli-python-3', 'All required sync blacklisted files are configured!'))
+        else:
+            self.failed.append(('cli-python-3', 'Blacklisted sync files section misses some required files!'))
+        return result
 
     def check_dependencies_not_outdated(self) -> bool:
         """
@@ -74,7 +93,7 @@ class CliPythonLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
                     latest_dependency_version = pip_dep_json['info']['version']
                     if parse_version(pip_dependency_version) < parse_version(latest_dependency_version):
                         self.warned.append(('cli-python-2', f'Version {pip_dependency_version} of {pip_dependency_name} is not the latest available: '
-                        f'{latest_dependency_version}'))  # noqa: E128
+                                                            f'{latest_dependency_version}'))  # noqa: E128
                 else:
                     self.failed.append(('cli-python-2', f'Could not find pip dependency using the PyPi API: {pip_dependency_name}=={pip_dependency_version}'))
 
@@ -136,6 +155,20 @@ class CliJavaLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
 
     def lint(self, skip_external):
         super().lint_project(self, self.methods)
+
+    def check_sync_section(self) -> bool:
+        """
+        Check the sync_files_blacklisted section containing every required file!
+        """
+        config_linter = ConfigLinter(f'{self.path}/cookietemple.cfg', self)
+        result = config_linter.check_section(section_items=config_linter.parser.items('sync_files_blacklisted'), section_name='sync_files_blacklisted',
+                                             main_linter=self, blacklisted_sync_files=[[('build_gradle', 'build.gradle'), ('changelog', 'CHANGELOG.rst')], -1],
+                                             error_code='cli-java-2', is_sublinter_calling=True)
+        if result:
+            self.passed.append(('cli-java-2', 'All required sync blacklisted files are configured!'))
+        else:
+            self.failed.append(('cli-java-2', 'Blacklisted sync files section misses some required files!'))
+        return result
 
     def java_files_exist(self) -> None:
         """