+++ cookietemple/lint/domains/lib.py
@@ -1,7 +1,7 @@
 import os
 from typing import List
 
-from cookietemple.lint.template_linter import TemplateLinter, files_exist_linting, GetLintingFunctionsMeta
+from cookietemple.lint.template_linter import TemplateLinter, files_exist_linting, GetLintingFunctionsMeta, ConfigLinter
 
 CWD = os.getcwd()
 
@@ -12,6 +12,20 @@ class LibCppLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
 
     def lint(self, skip_external):
         super().lint_project(self, self.methods)
+
+    def check_sync_section(self) -> bool:
+        """
+        Check the sync_files_blacklisted section containing every required file!
+        """
+        config_linter = ConfigLinter(f'{self.path}/cookietemple.cfg', self)
+        result = config_linter.check_section(section_items=config_linter.parser.items('sync_files_blacklisted'), section_name='sync_files_blacklisted',
+                                             main_linter=self, blacklisted_sync_files=[[('changelog', 'CHANGELOG.rst')], -1],
+                                             error_code='lib-cpp-2', is_sublinter_calling=True)
+        if result:
+            self.passed.append(('lib-cpp-2', 'All required sync blacklisted files are configured!'))
+        else:
+            self.failed.append(('lib-cpp-2', 'Blacklisted sync files section misses some required files!'))
+        return result
 
     def cpp_files_exist(self) -> None:
         """