+++ docs/conf.py
@@ -21,8 +21,8 @@ project = u'cookietemple'
 copyright = u'2020, Lukas Heumos, Philipp Ehmele, the cookiejar organization'
 author = u'Lukas Heumos, Philipp Ehmele, the cookiejar organization'
 
-version = '1.2.2-SNAPSHOT'
-release = '1.2.2-SNAPSHOT'
+version = '1.2.2'
+release = '1.2.2'
 
 language = None